--- src/common/operator/like_operators.cpp
@@ -1,47 +1,0 @@
-#include "duckdb/common/operator/like_operators.hpp"
-
-namespace duckdb {
-
-bool Like::Operation(const char *s, const char *pattern, const char *escape) {
-	const char *t, *p;
-
-	t = s;
-	for (p = pattern; *p && *t; p++) {
-		if (escape && *p == *escape) {
-			p++;
-			if (*p != *t) {
-				return false;
-			}
-			t++;
-		} else if (*p == '_') {
-			t++;
-		} else if (*p == '%') {
-			p++;
-			while (*p == '%') {
-				p++;
-			}
-			if (*p == 0) {
-				return true; /* tail is acceptable */
-			}
-			for (; *p && *t; t++) {
-				if (Like::Operation(t, p, escape)) {
-					return true;
-				}
-			}
-			if (*p == 0 && *t == 0) {
-				return true;
-			}
-			return false;
-		} else if (*p == *t) {
-			t++;
-		} else {
-			return false;
-		}
-	}
-	if (*p == '%' && *(p + 1) == 0) {
-		return true;
-	}
-	return *t == 0 && *p == 0;
-}
-
-} // namespace duckdb
