+++ src/common/vector_operations/CMakeLists.txt
@@ -9,7 +9,6 @@ add_library_unity(duckdb_vector_operations
                   generators.cpp
                   gather.cpp
                   hash.cpp
-                  like.cpp
                   null_operations.cpp
                   numeric_bitwise_operators.cpp
                   numeric_inplace_operators.cpp
