+++ src/function/scalar/string/like.cpp
@@ -7,21 +7,61 @@ using namespace std;
 
 namespace duckdb {
 
-static void like_function(ExpressionExecutor &exec, Vector inputs[], index_t input_count, BoundFunctionExpression &expr,
-                          Vector &result) {
-	result.Initialize(TypeId::BOOLEAN);
-	VectorOperations::Like(inputs[0], inputs[1], result);
-}
+struct LikeOperator {
+	static bool Operation(const char *left, const char *right, const char *escape = nullptr);
+};
 
-static void not_like_function(ExpressionExecutor &exec, Vector inputs[], index_t input_count,
-                              BoundFunctionExpression &expr, Vector &result) {
-	result.Initialize(TypeId::BOOLEAN);
-	VectorOperations::NotLike(inputs[0], inputs[1], result);
+struct NotLikeOperator {
+	static inline bool Operation(const char *left, const char *right, const char *escape = nullptr) {
+		return !LikeOperator::Operation(left, right, escape);
+	}
+};
+
+bool LikeOperator::Operation(const char *s, const char *pattern, const char *escape) {
+	const char *t, *p;
+
+	t = s;
+	for (p = pattern; *p && *t; p++) {
+		if (escape && *p == *escape) {
+			p++;
+			if (*p != *t) {
+				return false;
+			}
+			t++;
+		} else if (*p == '_') {
+			t++;
+		} else if (*p == '%') {
+			p++;
+			while (*p == '%') {
+				p++;
+			}
+			if (*p == 0) {
+				return true; /* tail is acceptable */
+			}
+			for (; *p && *t; t++) {
+				if (LikeOperator::Operation(t, p, escape)) {
+					return true;
+				}
+			}
+			if (*p == 0 && *t == 0) {
+				return true;
+			}
+			return false;
+		} else if (*p == *t) {
+			t++;
+		} else {
+			return false;
+		}
+	}
+	if (*p == '%' && *(p + 1) == 0) {
+		return true;
+	}
+	return *t == 0 && *p == 0;
 }
 
 void LikeFun::RegisterFunction(BuiltinFunctions &set) {
-	set.AddFunction(ScalarFunction("~~", {SQLType::VARCHAR, SQLType::VARCHAR}, SQLType::BOOLEAN, like_function));
-	set.AddFunction(ScalarFunction("!~~", {SQLType::VARCHAR, SQLType::VARCHAR}, SQLType::BOOLEAN, not_like_function));
+	set.AddFunction(ScalarFunction("~~", {SQLType::VARCHAR, SQLType::VARCHAR}, SQLType::BOOLEAN, ScalarFunction::BinaryFunction<const char*, const char*, bool, LikeOperator, true>));
+	set.AddFunction(ScalarFunction("!~~", {SQLType::VARCHAR, SQLType::VARCHAR}, SQLType::BOOLEAN, ScalarFunction::BinaryFunction<const char*, const char*, bool, NotLikeOperator, true>));
 }
 
 } // namespace duckdb
